[CC - 8] Add Custom Cell, Label, add Formatter
